logic/search-service/requirements.txt
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
PyJWT==2.6.0
Werkzeug==2.3.7
psycopg2-binary==2.9.3
SQLAlchemy==1.4.46
flask-cors==3.0.10
kafka-python==2.0.2
